Fix errors of test cases
